requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.7
django-cleanup==5.1.0
django-extensions==3.1.0
djangorestframework==3.12.2
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
Markdown==3.3.3
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
postgres==3.0.0
proto-plus==1.13.0
protobuf==3.15.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
